PROD-347: Make token endpoint backwards compatible with legacy roles
Roles that were created prior to the token endpoint will not have client
credentials in storage so fail the request and note that the role needs
to be updated or recreated.

For Datadog we really only currently have a single role which needs to
migrate so we can either update or recreate the role. But having the
option to update the role is nice so that things will automatically get
upgraded as they are touched.