PROD-599: Use the MS Graph API for atomic add/remove password operations
Azure Active Directory Graph API has been deprecated and does not
provide support for atomically creating/removing passwords on an
application. As a result, there is a race conditions that can occur when
creds are being created for a static application that is configured on
multiple mounts or across multiple Vault clusters.

Note that this changeset does not completely remove the Azure AD Graph
API client. Applications are still created/removed using the deprecated
API.

Additionally, although using the MS Graph API is a net benefit, it
itself has reliability issues when handling multiple requests in
parallel. More details can be found in
https://github.com/mdgreenfield/microsoft-graph-api-reliability